+++ src/operations/SemanticTokens.cpp
@@ -391,7 +391,7 @@ std::optional<lsp::SemanticTokens> WorkspaceFolder::semanticTokens(const lsp::Se
     auto moduleName = fileResolver.getModuleName(params.textDocument.uri);
     auto textDocument = fileResolver.getTextDocument(moduleName);
     if (!textDocument)
-        return std::nullopt;
+        throw JsonRpcException(lsp::ErrorCode::RequestFailed, "No managed text document for " + moduleName);
 
     // Run the type checker to ensure we are up to date
     if (frontend.isDirty(moduleName))