+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Fixed inlay hints no longer showing up
 - Fixed inlay hints not showing up in first load of file until a dummy change is made
+- Fixed DM types not generated for `script` nodes. Improved autocomplete will now be provided for non-DataModel projects (e.g. `Tool` as Root)
 
 ## [1.11.1] - 2022-10-01